Generalize binary tree to void * content
When working on projects that demand something
more complex than just simple arrays it might
be usefull to have a generalized content that can support
any type of structure.
The drawback is that now basic functions of the structure need
to be passed on when creating and manipulating the tree.